use-graph-controller as hook, wip, still need to type and swap passed params for ones we can get from context
